+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Dropped support for Twig 1
 * Removed `DoctrineCacheUAFamilyParser` (use `PsrCacheUAFamilyParser` instead)
 * All classes have been marked as `final`
+* Renamed `WhitelistBasedTargetValidator` class to `AllowListBasedTargetValidator`
 
 ### 2.12.0 (xxxx-xx-xx)
 * Filter moz-extension reports